Track and ignore reply_to message
Slack sends us the last message(s) we sent when re reconnet to the real-time api. I’m not sure why is does this, and I haven’t found a good explanation.

This fix treats the symptom, not the cause. It would be good to dig deeper into this sometime.