Refactor how context is used by states
A new feature to start a conversation with a slack user outside of the normal message loop highlighted an incompatibilty with how we’re extracting the channel from the context.

In the case of starting a new conversation, we don’t have a context. This refactor lays the ground work for a start conversation feature, without changing any existing functionality.