Periodically check and restart dead Slack RTC
We are seeing a problem where the Slack real-time client occasionally
disconnects silently. In this scenario the bot user goes offline but the
Ruby process doesn't crash and the RTC still seems to be connected.

As a workaround, this commit introduces a Ping class that can be used to
check wither the bot is online. It uses the Slack web client to do this
(the RTC does not have an API to check online status).

Then, when starting the bot, we first open the RTC and then enter a loop
where we use the Ping every minute. If the bot ever goes offline, we
stop the RTC and re-open it. When this happens we also send a message to
the logs and to the `warn_handler` (e.g. Rollbar).

This solution was inspired by this GitHub comment:
https://github.com/slack-ruby/slack-ruby-client/issues/208#issuecomment-419425356