Remove Rails cops from RuboCop config
Before, we were inheriting our rubocop config from the c5-conventions
repo. That config has changed over time to include things that are
specific to Rails. This makes sense, because 99% of the time when we are
using rubocop, it is on Rails projects. However this caused rubocop in
socrates to break: the c5-conventions assume the presence of the
`rubocop-rails` gem, which socrates does not have.

```
$ bundle exec rubocop
cannot load such file -- rubocop-rails
```

Fix by copying the c5-conventions rubocop.yml directly into this
project, then manually modify to remove the Rails-specific stuff.

Note that since we are no longer inheriting from c5-conventions, this
copied config might "drift" or become out of date compared to the latest
c5-conventions. I feel like this is an acceptable compromise because
overhauling the conventions themselves to accommodate non-Rails projects
(e.g. splitting into two files, Rails and non-Rails) seems like a lot of
work and ongoing maintenance in order to support the 1% of projects like
socrates that would benefit.