Update to latest rubocop gems
By default, rubocop makes you explicitly opt in or out of new rules that
are added in each release. Otherwise it loudly prints warnings to the
console. This is annoying and makes updating gems tedious. In practice,
we almost always end up enabling the new rules.

This commit updates our rubocop config with:

```
NewCops: enable
```

This automatically opts us into any new rules that are added. That means
there is no need to explicitly enable individual rules, and so our
rubocop config becomes much smaller.

This commit updates rubocop and rubocop-performance to their latest
versions and thus enables all the new rules that have been added in
those versions. The only rule that caused violations (and that I have
disabled), is `Style/AccessorGrouping`.